separate interpret state into own file
