+++ cilantro/nodes/delegate/__init__.py
@@ -1,2 +1,3 @@
 from cilantro.nodes.delegate.delegate import DelegateBootState
 from cilantro.nodes.delegate.consensus import DelegateConsensusState
+from cilantro.nodes.delegate.interpret import DelegateInterpretState
